Add: isVR option + Clean: refactor AbstractBrushStroke to ES6
